refactor(connection settings): move selector
