+++ src/ui/units/dash/containers/Dialogs/Control2/Sections/CommonSettingsSection/ConnectionSettings/components/ConnectionSelector/ConnectionSelector.tsx
@@ -4,9 +4,9 @@ import {useDispatch, useSelector} from 'react-redux';
 import {EntryScope} from 'shared';
 import type {GetEntryResponse} from 'shared/schema';
 
-import {setSelectorDialogItem} from '../../../../../../store/actions/dashTyped';
-import {selectSelectorDialog} from '../../../../../../store/selectors/dashTypedSelectors';
-import {EntrySelector} from '../EntrySelector/EntrySelector';
+import {setSelectorDialogItem} from '../../../../../../../../store/actions/dashTyped';
+import {selectSelectorDialog} from '../../../../../../../../store/selectors/dashTypedSelectors';
+import {EntrySelector} from '../../../EntrySelector/EntrySelector';
 
 const i18nConnectionBasedControlFake = (str: string) => str;
 const getConnectionLink = (connectionId: string) => `/connections/${connectionId}`;
